Added modifiers

+++ Backend/StoneActionServer.DAL/Models/Crafting/CraftingRecipe.cs
@@ -5,7 +5,7 @@ public class CraftingRecipe
     public int Id { get; set; }
     public int ResultItemId { get; set; }
     public int ResultQuantity { get; set; }
-    public double ChanceOfSuccess { get; set; }
+    public float ChanceOfSuccess { get; set; }
     public int CraftingTimeSeconds { get; set; }
     public int CraftingType { get; set; }
